Moved csv parsing into a separate utility file
